Condence common code for cm scenario
